Gutenberg metadata about each work

Selected fields of metadata about each of the Project Gutenberg works. These were collected using the gitenberg Python package, particularly the pg_rdf_to_json function.

gutenberg_metadata

Format

A tbl_df (see tibble or dplyr) with one row for each work in Project Gutenberg and the following columns:

gutenberg_id

Numeric ID, used to retrieve works from Project Gutenberg

title

Title

author

Author, if a single one given. Given as last name first (e.g. "Doyle, Arthur Conan")

author_id

Project Gutenberg author ID

language

Language ISO 639 code, separated by / if multiple. Two letter code if one exists, otherwise three letter. See https://en.wikipedia.org/wiki/List_of_ISO_639-2_codes

gutenberg_bookshelf

Which collection or collections this is found in, separated by / if multiple

rights

Generally one of three options: "Public domain in the USA." (the most common by far), "Copyrighted. Read the copyright notice inside this book for details.", or "None"

has_text

Whether there is a file containing digits followed by .txt in Project Gutenberg for this record (as opposed to, for example, audiobooks). If not, cannot be retrieved with gutenberg_download

Details

To find the date on which this metadata was last updated, run attr(gutenberg_metadata, "date_updated").

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(stringr)

gutenberg_metadata

#> # A tibble: 51,997 x 8
#>    gutenberg_id title  author  gutenberg_autho… language gutenberg_books… rights
#>           <int> <chr>  <chr>              <int> <chr>    <chr>            <chr> 
#>  1            0  NA    NA                    NA en       NA               Publi…
#>  2            1 "The … Jeffer…             1638 en       United States L… Publi…
#>  3            2 "The … United…                1 en       American Revolu… Publi…
#>  4            3 "John… Kenned…             1666 en       NA               Publi…
#>  5            4 "Linc… Lincol…                3 en       US Civil War     Publi…
#>  6            5 "The … United…                1 en       American Revolu… Publi…
#>  7            6 "Give… Henry,…                4 en       American Revolu… Publi…
#>  8            7 "The … NA                    NA en       NA               Publi…
#>  9            8 "Abra… Lincol…                3 en       US Civil War     Publi…
#> 10            9 "Abra… Lincol…                3 en       US Civil War     Publi…
#> # … with 51,987 more rows, and 1 more variable: has_text <lgl>

gutenberg_metadata %>%
  count(author, sort = TRUE)

#> # A tibble: 16,190 x 2
#>    author                                  n
#>    <chr>                               <int>
#>  1 NA                                   3457
#>  2 Various                              3199
#>  3 Anonymous                             724
#>  4 Shakespeare, William                  317
#>  5 Lytton, Edward Bulwer Lytton, Baron   219
#>  6 Twain, Mark                           214
#>  7 Ebers, Georg                          171
#>  8 Dickens, Charles                      163
#>  9 Verne, Jules                          151
#> 10 Unknown                               141
#> # … with 16,180 more rows

# look for Shakespeare, excluding collections (containing "Works") and translations
shakespeare_metadata <- gutenberg_metadata %>%
  filter(author == "Shakespeare, William",
         language == "en",
         !str_detect(title, "Works"),
         has_text,
         !str_detect(rights, "Copyright")) %>%
         distinct(title)

if (FALSE) {
shakespeare_works <- gutenberg_download(shakespeare_metadata$gutenberg_id)
}

# note that the gutenberg_works() function filters for English
# non-copyrighted works and does de-duplication by default:

shakespeare_metadata2 <- gutenberg_works(author == "Shakespeare, William",
                                         !str_detect(title, "Works"))

# date last updated
attr(gutenberg_metadata, "date_updated")

#> [1] "2016-05-05"